# cmd/docs/swagger.yaml
definitions:
...
  aqary_admin_internal_domain_requests_user.CheckEmailRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  aqary_admin_internal_domain_requests_user.CheckUsernameRequest:
    properties:
      username:
        type: string
    required:
    - username
    type: object
...
  aqary_admin_internal_domain_requests_user.UpdateUserRequest:
    properties:
      city:
        type: integer
      community:
        type: integer
      company_id:
        type: integer
      country:
        type: integer
      department:
        type: integer
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      permissions:
        type: string
      phone_number:
        type: string
      roles_id:
        type: integer
      state:
        type: integer
      sub_community:
        type: integer
    type: object
...
  aqary_admin_internal_domain_responses_user.AllUserOutput:
    properties:
      city:
        type: string
      community:
        type: string
      country:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      location:
        type: string
      phone:
        type: string
      role:
        type: string
      state:
        type: string
      status:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      sub_community:
        type: string
      username:
        type: string
    type: object
...
  aqary_admin_internal_domain_responses_user.DeletedAqaryUserOutput:
    properties:
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
      role:
        type: string
    type: object
...
  aqary_admin_internal_domain_responses_user.UpdatedUserOutput:
    properties:
      city:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      community:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      country:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      department:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      permission:
        items:
          type: integer
        type: array
      phone:
        type: string
      role:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      state:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      sub_community:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      sub_section_permission:
        items:
          type: integer
        type: array
      username:
        type: string
    type: object
...
  aqary_admin_internal_domain_responses_user.UserOutput:
    properties:
      city:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      community:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      country:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      department: {}
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      permission: {}
      phone:
        type: string
      role: {}
      state:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      sub_community:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      user_type_id:
        type: integer
      username:
        type: string
    type: object
...
  aqary_admin_internal_domain_sqlc_sqlc.SearchAllAgentRow:
    properties:
      id:
        type: integer
      phone_number:
        $ref: '#/definitions/pgtype.Text'
      username:
        type: string
    type: object
...
  aqary_admin_pkg_utils.ErrResponseSwagger:
    properties:
      error:
        example: any error message
        type: string
    type: object
...
  aqary_admin_pkg_utils_fn.CustomFormat:
    properties:
      id:
        type: integer
      label:
        type: string
    type: object
...
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Testing Swagger APIs.
  title: Aqary  APIs
  version: "1.0"
paths:
  /api/agents/search:
    get:
      consumes:
      - application/json
      description: Search for agents based on a search string
      parameters:
      - description: Search string
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.SearchAllAgentRow'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Search all agents
      tags:
      - agents
...
  /api/user/checkEmail:
    post:
      description: Check if an email already exists
      parameters:
      - description: CheckEmailRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.CheckEmailRequest'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: CheckEmail
      tags:
      - users
  /api/user/checkUsername:
    post:
      description: Check if a username already exists
      parameters:
      - description: CheckUsernameRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.CheckUsernameRequest'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: CheckUsername
      tags:
      - users
...
  /api/user/deleteAqaryUser/{id}:
    delete:
      description: Delete an Aqary user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Delete an Aqary user
      tags:
      - aqary_users
...
  /api/user/getAllAqaryUsers:
    get:
      description: Get a list of all Aqary users with pagination
      parameters:
      - description: Page Size
        in: query
        name: page_size
        required: true
        type: integer
      - description: Page Number
        in: query
        name: page_no
        required: true
        type: integer
      - description: Search string
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_responses_user.AllUserOutput'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Get all Aqary users
      tags:
      - aqary_users
  /api/user/getAllAqaryUsersByCountryId:
    get:
      description: Get a list of all Aqary users filtered by country with pagination
      parameters:
      - description: Page Size
        in: query
        name: page_size
        required: true
        type: integer
      - description: Page Number
        in: query
        name: page_no
        required: true
        type: integer
      - description: Country ID
        in: query
        name: country_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_responses_user.AllUserOutput'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Get all Aqary users by country
      tags:
      - aqary_users
  /api/user/getAllDeletedAqaryUser:
    get:
      description: Get a list of all deleted Aqary users with pagination
      parameters:
      - description: Page Size
        in: query
        name: page_size
        required: true
        type: integer
      - description: Page Number
        in: query
        name: page_no
        required: true
        type: integer
      - description: Search string
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_responses_user.DeletedAqaryUserOutput'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Get all deleted Aqary users
      tags:
      - aqary_users
  /api/user/getAllDeletedAqaryUserWithoutPagination:
    get:
      description: Get a list of all deleted Aqary users without pagination
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_responses_user.DeletedAqaryUserOutput'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Get all deleted Aqary users without pagination
      tags:
      - aqary_users
...
  /api/user/getSingleAqaryUser/{id}:
    get:
      description: Get detailed information of a single Aqary user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_responses_user.UserOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Get detailed information of a single Aqary user
      tags:
      - aqary_users
...
  /api/user/restoreAqaryUser/{id}:
    post:
      description: Restore a previously deleted Aqary user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Restore a deleted Aqary user
      tags:
      - aqary_users
  /api/user/updateAqaryUser/{id}:
    put:
      description: Update details of an Aqary user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User details to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.UpdateUserRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_responses_user.UpdatedUserOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Update an Aqary user
      tags:
      - aqary_users
...
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"